+++ packages/rspack-config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ice/rspack-config
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [78f850fa]
+  - @ice/shared-config@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes